Move enumerate_instance_version to Entry
Also adds `EntryV1_1` trait

# Conflicts:
#	ash/src/version.rs
