+++ ash/src/instance.rs
@@ -86,15 +86,6 @@ impl<V: FunctionPointers> Instance<V> {
 pub trait InstanceV1_1: InstanceV1_0 {
     fn fp_v1_1(&self) -> &vk::InstanceFnV1_1;
 
-    unsafe fn enumerate_instance_version(&self) -> VkResult<vk::uint32_t> {
-        let mut api_version = mem::uninitialized();
-        let err_code = self.fp_v1_1().enumerate_instance_version(&mut api_version);
-        match err_code {
-            vk::Result::SUCCESS => Ok(api_version),
-            _ => Err(err_code),
-        }
-    }
-
     fn enumerate_physical_device_groups(&self) -> VkResult<Vec<vk::PhysicalDeviceGroupProperties>> {
         unsafe {
             let mut group_count = mem::uninitialized();
