+++ ash/src/version.rs
@@ -1,5 +1,5 @@
 pub use device::{DeviceV1_0, DeviceV1_1};
-pub use entry::EntryV1_0;
+pub use entry::{EntryV1_0, EntryV1_1};
 pub use instance::{InstanceV1_0, InstanceV1_1};
 use std::mem;
 use vk;
